Let infer_nrows default to None to keep behavior
Rework so that if infer_nrows is not set, detect_colspecs still uses
n=100, preserving previous behavior. If infer_nrows comes down the pipe,
the default n is overwritten. Also make sure that the new parameter
infer_nrows is the trailing keyword in all function calls.

Fix the documentation to reflect these changes.