DOC: Generalize documentation of rename function
In this commit, the documentation of the rename function of the NDFrame,
is generalized over all its children using the function: DataFrame,
Series and Panel. These three classes now use the Substitute decorator
to add the generic part of the documentation to their own, using the
same decorator to replace class specific parameters into the generic
documentation.