DOC: Fixed redirects in various parts of the documentation
Links in the docs to part way down a page were being sent to the top of the page due to a generic redirect from pandas.pydata.org/pandas-doc/stable > pandas.pydata.org/pandas-doc/stable/user_guide. These links have been change to go straight to the .../user_guide/... version of the page to ensure the link goes to the correct part of the docs.