Follow example of forking doc from get_loc method
Originally had made a template from the documentation so a common basis could be
used with the IntervalIndex adding a warning that the other methods were as yet
unsupported. In the end because I am not clear on what effect putting a template
in the core documentation dictionary I feel back to just following get_loc and
forking the doc string.