catch iterator input in io.json.json_normalize
The data argument in json_normalize should be an dictionary or list of dictionaries. If an iterator is passed, unexpected behavior can occur. In particular, this will result in the loss of the first row. See https://stackoverflow.com/questions/56362810/missing-first-document-when-loading-multi-document-yaml-file-in-pandas-dataframe.

In my proposal, an error is thrown. Alternatively, the iterator could be turned into a list:

```
if data is iter(data):
    data = [x for x in data]
```