BUG: XlsxWriter ignoring formats on numpy types if merged cells
The write_cells method takes a 'cells' parameter that is itself derived
from a generator of the various cells in the dataframe. It calls
_value_with_format to convert any numpy types to Python types for the
Excel writer. In the section of code that deals with writing merged
cells, the original 'cell.val' parameter was being passed into the
writer, rather than the 'val' returned from the format function. This
caused incompatible numpy or Pandas formats to get passed into the
writer when they were in a 'merged' (grouped) DataFrame cell. In my
case I had a Period object in the DataFrame index. All that needed to
be done was simply removing 'cell.'