Make a fastpath conditional on `data is None`
As per the pre-existing comment in the code: this is a fastpath for the
case `data is None`. Better include an if-statement to make sure this
condition actually holds.

When this if-statement is not present, failure occurs at a later point
in the code when an empty index is used, because that later code expects data
to be array-like, but we convert to a scalar here.

fix #26469