Rewrite Syntax: python explode.py infile template [range]
The template argument is used to construct the names of the
individual frame files.  The frames are numbered file001.ext,
file002.ext, etc.  You can insert %d to control the placement
and syntax of the frame number.

The optional range argument specifies which frames to extract.
You can give one or more ranges like 1-10, 5, -15 etc.  If
omitted, all frames are extracted. method of frame so it doesn't require exploding a series twice