PERF: data = np.nan to speed up empty dataframe creation (GH1234)
Check that we have enough informations to know the shape of the final
dataframe and set the value of data to np.nan if not set.
With the former behaviour data is defaults always to None,
hence the output dataframe has dtypes object and its creation
is slow.
With the proposed code the output dataframe has dtypes float and is
faster to create but only when columns and index are given.