Don't fail when plotting Series/DataFrame with no row
When Series or DataFrames was empty (no cells) after removing columns
without a suitable type (.select_dtypes), pandas was throwing a "no
numeric data to plot" exception.

This can happen:

1) either because there is no column with a suitable type;

2) or if there is no row.

Raising an exception in the first case makes sense but we should
probably avoid throwing an exception in the second case.