Fix a typo in "computation.rst" in document.
There's `np.random.np.random` in /doc/source/user_guide/computation.rst, which I believe is a typo. But the weird thing is there's actually `np.random.np` in numpy (1.16.4), but not in numpy (1.17.2). That's maybe why the doc build passed before. While I'm trying to build the doc locally with numpy 1.17.2, it failed.