Add built-in funcion for Styler to format the text displayed for missing values
As described in GH #28358, user who wants to control how NA values are printed
while applying styles to the output will have to implement their own formatter.
(so that the underlying data will not change and can be used for styling)

Since the behavior is common in styling (for reports etc.), suggest to add this
shortcut function to enable users format their NA values as something like '--'
or 'Not Available' easily.

example usage: `df.style.highlight_max().format_null('--')`