[WIP]: Indexing with BooleanArray
This implements indexing a Series with a BooleanArray for some
dtypes. NA values in the mask propagate.

This required updating each EA's `__getitem__` to correctly propagate
NA. This isn't an option for ndarray-backed Series, which don't
understand BooleanArray.

No need to review in detail. The primary motivation for pushing this up
is to better understand what indexing with NA values will look like in
practice.