DOC: Fixed contributors for bugfix releases
Previously we showed just contributors who manually backported commits
to the maintenance branch.

```
Contributors
============

A total of 8 people contributed patches to this release.  People with a
"+" by their names contributed a patch for the first time.

* Daniel Saxton
* Joris Van den Bossche
* MeeseeksMachine
* MomIsBestFriend
* Pandas Development Team
* Simon Hawkins
* Tom Augspurger
* jbrockmendel
```

Fixed

```
Contributors
============

A total of 22 people contributed patches to this release.  People with a
"+" by their names contributed a patch for the first time.

* Anna Daglis +
* Daniel Saxton
* Irv Lustig
* Jan Škoda +
* Joris Van den Bossche
* Justin Zheng +
* Kaiqi Dong
* Kendall Masse +
* Marco Gorelli
* Matthew Roeschke +
* Pedro Reys +
* Prakhar Pandey +
* Robert de Vries +
* Rushabh Vasani +
* Simon Hawkins +
* Stijn Van Hoey +
* Terji Petersen +
* Tom Augspurger
* William Ayd
* alimcmaster1
* gfyoung +
* jbrockmendel
```