API/TST: Call __finalize__ in more places
Progress towards https://github.com/pandas-dev/pandas/issues/28283. This
adds tests that ensures `NDFrame.__finalize__` is called in more places.
Thus far I've added tests for anything that meets the following rule:

> Pandas calls `NDFrame.__finalize__` on any NDFrame method that returns
> another NDFrame.

I think that given the generality of `__finalize__`, making any kind of
list of which methods should call it is going to be somewhat arbitrary.
That rule errs on the side of calling it too often, which I think is my
preference.