Solve missing interpolation method (cubicspline)
By commit 8bb2cc1 scipy.interpolate.CubicSpline method is
referenced in the pandas documentation (see pandas/core/generic.py)
but it is not wrapped by any interpolation method. This commit
solves this adding the corresponding wrapper. SciPy's CubicSpline
is a cubic spline data interpolator that allows explicit control
of the boundary conditions for the interval.

Changes to be committed:
	modified:   ../../../core/missing.py
	modified:   test_interpolate.py