typo: rows -> columns
I'm new to pandas and reading the docs for the first time.  By my reading of the first part of this paragraph, there's a dictionary of lists and the lists are `["Braund...", "Allen...", ...]`, `[22, 35, 58]`, and `["male", "male", "female"]`.

The original sentence says: "... and the values in each list as rows of the DataFrame", but the values in the first list (`["Baund...]`) are the values in the first **column** of the illustration.  Likewas the list `[22,...]` is the second column and `["male",...]` is the third.