DOC: correction to "size" in the plotting.bootstrap_plot docstring
The documentation for plotting.bootstrap_plot says that the size argument must be greater than or equal to the length of the series. Bootstrapping typically uses subsamples of a size less than or equal to the size of the data set, so I corrected the docs accordingly. Note that making size greater than the length of the series throws the following ValueError: Sample larger than population or is negative