DOC: Rst Formatting, make sure continuation prompt are used.
Without the `...` for some tools it may look like
`with ExcelWriter('path_to_file.xlsx',` is an input and
`date...` is the output.

I initially thought that I could use the hanging indent to separate
code continuation from output but `pandas/core/indexing.py`
proved me wrong as the example just after the one I fixed have
the first line of the output be indented.