TST: don't expect same indexer for Index and PeriodIndex
The new sort_values implementation for Index is stable when sorting in
descending order. However, we expect PeriodIndex.sort_values to be
unstable (and have tests for this). So until we enforce the same
convention for sorting duplicate values in descending order, we
can't check that the indexers for these classes are the same.