Revert __finalize__ call in DataSplitter
The call was an attempt to ensure that `__finalize__` was called for DataFrames returned when iterating over grouping result, but it caused unreasonable performance degradation.
The relevant test has been marked with `xfail` in order to preserve it. When a solution to the issue is implemented, this test should be reactivated.