REGR: Allow positional arguments in DataFrame.agg
Although `DataFrame.agg` is documented as accepting `func, axis, *args, **kwargs`
with `*args` and `**kwargs` passed to `func`, passing positional
arguments raises a TypeError.

The reason for this is that the internal call to `self._aggregate` uses
a keyword argument (axis) before passing *args and `**kwargs`, and as such the
first positional argument is always interpreted as a second
specification of `axis`, which raises TypeError.

Prior to commit 433c9007781080658553fbef1a4d0c2813b404c0, TypeErrors
were being suppressed, falling back to `self.apply`, which in v1.1.0
turned into an error.

This fixes issue GH-36948.