BUG: GH37635
Previously, the variable "cols" was created solely to concatenate
the lists "index" and "columns" when the "values" parameter was
left as None.  It did so in a way that it modified the variable
passed as "index" using list.extend(), affecting the caller's
namespace.

This change simply uses an anonymous list in place of the
variable "cols"