Fixes for combining axes when non-intersecting labels are duplicated
This fixes a previous issue where `get_objs_combined_axis` would not throw and error for duplicated indices in all cases.
Now we check to make sure that the duplicated values don't show up in the intersection.
This allows safe indexing when there are duplicated values, but they don't show up in the intersection.