Allow passing custom sentinel into MultiIndex.format
Commit b0468aa45f3 seems to have unintentionally removed functionality
to pass custom sentinel values for sparsifying in MultiIndex.format().
Previously, any "Truthy" value other than 1 and True was used as the
sentinel, which is useful particularly with adjoin=False. The current
logic does not make sense: it checks sparsify for "Falsey" values under
"if sparsify:" condition. Under this if sparsify cannot be Falsey.

Also, add an explicit sentinel argument to make it less confusing.