ENH: Allow bar and barh plots to use index values for x position.
The bar plots X-axis ignores the current index values for the position,
so the bars are always at position 0, 1, 2... despite the index values.

This patch adds x_mode, a new parameter with three possible values:
- sequential: This is the current behaviour of bar and the default value.
- numerical: The X axis positions will be related to the real index values.
- categorical: The X axis tics includes all the possible categories,
even if are not all used in the index.

With these modes it is possible to create richer bar plots,
specially useful to compare results from different samples.

Now with blake and flake8 checked.