remove tensorflow-addons dependency. Remove FScore metric from NNInterface
remove tensorflow-addons dependency. Remove FScore metric from NNInterface
plitted the pandas plotting module into a general plotting framework able to call different backends and the current matplotlib backends. The idea is that other backends can be implemented in a simpler way, and be used with a common API by pandas users.

The API defined by the current matplotlib backend includes the objects listed next, but this API can probably be simplified. Here is the list with questions/proposals: