BUG: fix Frame getitem when column keys are nested tuples #43780
Multi-column selection when the column keys are nested tuples (for
example when a MultiIndex has at each level a tuple) raised error due
to a condition in asarray_tuplesafe.

A key list of form [((1,))] converts to [[[1]]] through the call to
np.array, resulting in ndim>2. However, only the case ndim=2 was
further handled, causing a multi-dimensional array to be returned and
subsequently raising error.

This commit extends the ndim=2 condition for these type of keys.