COMPAT: Python 3.9.8
Please see bpo 20028. https://github.com/python/cpython/pull/28705. 
This shows up in nightlies, but I have no way to test this. It should show up pretty soon in regular CI. 
Can someone with with the perms on the MacPython repo please trigger the nightlies manually to see if that fixes the builds? Thanks.