BUG: CategoricalIndex.get_indexer issue with NaNs (#45361) (#45373)
* BUG: CategoricalIndex.get_indexer issue with NaNs (#45361)

categorical_index_obj.get_indexer(target) yields incorrect results when
categorical_index_obj contains NaNs, and target does not. The reason
for this is that, if target contains elements which do not match any
category in categorical_index_obj, they are replaced by NaNs. In such
a situation, if categorical_index_obj also has NaNs, then the corresp
elements in target are mapped to an index which is not -1

eg:
ci = pd.CategoricalIndex([1, 2, np.nan, 3])
other = pd.Index([2, 3, 4])

ci.get_indexer(other)
In the implementation of get_indexer, other becomes [2, 3, NaN]
which is mapped to index 2, in ci

* BUG: CategoricalIndex.get_indexer issue with NaNs (#45361)

categorical_index_obj.get_indexer(target) yields incorrect results when
categorical_index_obj contains NaNs, and target does not. The reason
for this is that, if target contains elements which do not match any
category in categorical_index_obj, they are replaced by NaNs. In such
a situation, if categorical_index_obj also has NaNs, then the corresp
elements in target are mapped to an index which is not -1

eg:
ci = pd.CategoricalIndex([1, 2, np.nan, 3])
other = pd.Index([2, 3, 4])

ci.get_indexer(other)
In the implementation of get_indexer, other becomes [2, 3, NaN]
which is mapped to index 2, in ci

Update:
np.isnan(target) was breaking the existing codebase.
As a solution, I have enclosed this line in a try-except block

* BUG: CategoricalIndex.get_indexer issue with NaNs (#45361)

Update: Replaced `np.isnan` with `pandas.core.dtypes.missing.isna`

* Added a testcase to verify output behaviour

* Made pre-commit changes

* Added a test case without NaNs

* Moved NaN test to avoid unnecessary execution

* Re-aligned test cases

* Removed try-except block

* Cleaned up base.py

* Add GH#45361 comment to code

* Added whatsnew entry

* Resolved merge conflict

* Moved whatsnew entry to indexing section