TYP: Fix annotation of Grouper._set_grouper
The return value is never used. Don't pretend
that it is.

Ideally we don't even hold this state inside the Grouper, it would make
it easier to reason about. But just do this tiny change for now,
still an improvement.
See https://github.com/pandas-dev/pandas/pull/46258 for more details.