BUG: Fix infinite recursion loop when pivot of IntervalTree is ±inf
Attempts to fix #46658.

When the pivot of an IntervalTree becomes ±inf the construction of the
IntervalTree comes to an infinite loop recursion.  This patch tries to fix that
by catching those cases and set the pivot to a reasonable value.

Note that the tests are skipped on 32-bit systems (see #23440)