Add "rows" as option to "axis" docstring
The "axis" parameter of `pd.concat()` can take the string `'rows'` as an optional way of explicitly defining the default row axis for concatenation:

```
import pandas as pd

pd.concat([pd.DataFrame(), pd.DataFrame()], axis="rows")
```

Using `'rows'` instead of `'index'` or `0` might be more intuitive as a counterpart to `'columns'`. In any case, it's a possible option, so it might be helpful to represent that in the docstring.