Add "rows" as option to "axis" docstring
The "axis" parameter of `pd.concat()` can take the string `'rows'` as an optional way of explicitly defining the default row axis for concatenation:

```
import pandas as pd

pd.concat([pd.DataFrame(), pd.DataFrame()], axis="rows")
```

Using `'rows'` instead of `'index'` or `0` might be more intuitive as a counterpart to `'columns'`. In any case, it's a possible option, so it might be helpful to represent that in the docstring.

The proposed update in the docstring uses the suggested format based on pandas docstring guide:
https://pandas.pydata.org/docs/dev/development/contributing_docstring.html#parameter-types

>For axis, the convention is to use something like:
>axis : {0 or ‘index’, 1 or ‘columns’, None}, default None