PERF: Remove unnecessary(?) asof join functions
Several functions in the join cython are module are basically
just calling others, and can probably be removed. This wouldn't
be a big deal inside python, but in this case it cuts down
considerably on the number of generated cython code.

The tests pass for me locally when I do this and there seems
to be significant space savings. Will provide more detail after
I see CI passing (want to make sure I'm not missing something
first)