Adds Hamilton as a development tool
Hamilton is a paradigm that helps one manage a pandas code base in an opinionated manner.

We have had great success with it at Stitch Fix, and it's got a growing open source
following. I think it helps one to write pandas in a way that gives
confidence if you want to run Pandas in production jobs.