Fix 48224: DataFrame.at must fail when adding more than one value
The problem occurred because when adding a new column to a DataFrame using
DataFrame.at, DataFrame._set_value falls back to DataFrame.loc before the
InvalidIndexError is raised.

The proposed solution makes sure that both, the column index and the row index
are determined so that both have a chance to raise.