Reword DataFrame.assign
I think wording is now a bit more precise, and the explicit
comparison with assignment helps to understand what this does.
I've also moved the note on multiple assignments into the **kargs
description because I think it really belongs there and is not
only a remark we should make further down in the docstring.